Adds wishlist feature
Implements a wishlist feature allowing users to save and track desired cards.

Introduces a new "Wishlist" component with similar search and filter functionalities as the binder.

Creates a "CardCollectionWish" component to display the wishlist and manage card counts.

Stores wishlist data in local storage for persistence.

Adds a Wishlist link to the Navbar.